fix based on comments
