+++ functionary/prompt_template/prompt_template_v2.py
@@ -562,36 +562,14 @@ class PromptTemplateV2(PromptTemplate):
     def get_force_function_call_prefix(self, function_name: str):
         return f"{function_name}{self.fn_param_sep_token}"
 
-    def get_raw_response_from_assistant_messages(
-        self,
-        messages: List[Dict[str, str]],
-        tool_func_choice: Union[str, Tool, Function],
-        default_tool_call_name: str,
-    ):
-        # Form raw response from messages list
-        raw_response = "".join(
-            [self.convert_message_to_prompt(message) for message in messages]
-        ).rstrip()
+    def get_raw_response_output(self, raw_response: str):
+        """Get raw response by removing null_content from the beginnging of raw response
+        (<|from|>assistant\n<|recipient|>)
 
-        # Remove null_response "<|from|>assistant\n<|recipient|>" from the beginning
-        null_response = self.convert_message_to_prompt({"role": "assistant"})
-        raw_response = raw_response[len(null_response) :]
+        Args:
+            raw_response (str): raw response with null_content still in it
+        """
 
-        # Remove stop tokens
-        for stop_token in self.get_stop_tokens_for_generation():
-            raw_response = raw_response.replace(stop_token, "")
+        null_content = self.convert_message_to_prompt({"role": "assistant"})
 
-        if tool_func_choice == "none":
-            raw_response = raw_response[len(self.get_force_text_generation_prefix()) :]
-        elif isinstance(tool_func_choice, Tool) or isinstance(
-            tool_func_choice, Function
-        ):
-            raw_response = raw_response[
-                len(
-                    self.get_force_function_call_prefix(
-                        function_name=default_tool_call_name
-                    )
-                ) :
-            ]
-
-        return raw_response
+        return raw_response[len(null_content) :]
